some updates
split fetchData -> group & user
