+++ Frontend/src/utils/fetchData/teacher/user.js
@@ -1,0 +1,16 @@
+import axios from '../../../axios'
+
+import { getUserToken } from '../../userTokenManager'
+
+export const fetchUsers = async () => {
+  try {
+    const { data } = await axios.get('/user/all', {
+      headers: {
+        Authorization: await getUserToken(),
+      },
+    })
+    return data
+  } catch (error) {
+    console.log(error)
+  }
+}
